+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Changed
+
+- Disable per-player tech owing to on-going issues pending the next game patch
+
 ## v5.87.1 - 2026-06-26
 
 ### Bugfix